Code/requirements.txt
alt_model_checkpoint==2.0.3
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.1
Pillow==9.5.0
tensorflow==2.12.0
tensorflow_addons==0.20.0
tensorflow_datasets==4.9.2
